create api.js
